Eye Form and GFS Update
Ready for 5.0.2
Moved PCP/ref doc into Communication engine along w/ Pharmacy data(+selector) + links to eRx in menubar and comm engine area as users requested.

If not digits(words are entered, the value defaults to “” now for the Glaucoma Flow Sheet.
It was failing for “U/A” etc for example post-enucleation.

some edits by bradymiller