Fixed a MySQL Innodb table lock bug (#4493)
If an exception is thrown in generating the UUIDs the tables were being
locked through an open transaction.  PHP did not have the Exception
class imported so SqlQueryException was not being caught, fatal error
occurred and the transaction was left open.

Fixing the import now lets the transaction be caught and then
rollbacked.