Openemr fhir encounters (#4550)
* Comment on unclear code.

* Formatting encounter to match ONC requirements.

Still need to fix location and see what we do with ONC dispatchReason
for encounters.

* Add Facility Location Reference for encounter

* Implement permission issues on location resource.

Added a type to the location query so we can restrict location contact
information based on the user's permission level.

* Fix location query for fhir resource.

* Support Discharge Disposition field for encounters.

FHIR requires that we support the ability for an encounter to have
marked what the discharge disposition of the patient was when they left
an encounter.  Most of our user base will have Home be the default, but
other types of discharges are now supported via the dropdown.

* Fix single Location resource endpoint

* Style fixes

* Add discharge_disposition to new install...

You'd think I'd get this right eventually... forgot to add the column
for a new database installation.

* Fix the dang comma

* Fix use statement and escape text/attrs

* Fix location security issues for patient access.

Had a bug where patients were still getting access to address data that
was not their own data.  Now clinician and patient home address
information is hidden from the patient.

* Fix styles.