Issue 4469 (#16) (#4505)
* Issue 4469 (#16)

This adds Audit Log logging to the patient-merge process.  Keeps track of which user did the merge, which patients were merged and the actions performed delete/update/etc on which table. (It does not log the raw SQL).

Note: I keep seeing a fail in a specific test, but not due to any of the code ... it fails looking for phpunit (which I see in the filesystem) it is linked to another spot, and I assume that is the problem- the link is not being created, or not being followed.  I've re-run the tests a couple of times, and it seems consistent.


* #4469 Patient Merge Audit Log

Added direct event audit lines for each merge event.

* #4469

Minor cleanup of comments etc.

* #4469 Missed log spots.

Missed some spots where logging should occur. Added those spots in.
Added $target_pid to deleteRows for passing that down to the log method.

* #4469 Moved logMergeEvent

Didn't like it being mashed in between the other methods,
moved it down to the bottom.

* #4469

Added use line for EventAuditLogger.

* #4469

Changed patient-merge to patient_merge
patient-merge gets a bit mangled when querying the log
for the 'event-type' ... the '-' is used to trim the last part off.

* #4469

Messing with the log messages, to make sure they read logically.

* #4469

Added additional tagging.

* #4469

Massive code formatting, for lint checking.

* #4469 : Formatting.

Ah, finally found the right format specification (ci/phpcs.xml)

Co-authored-by: Stuart <geektoast@gmail.com>

* put back line in header, put tranlation string on one line

Co-authored-by: Stuart <geektoast@gmail.com>