Fixes #4658 change uuid mapping to be on save. (#4679)
* Fixes #4658 change uuid mapping to be on save.

Created a FHIR module that listens to the vitals, patient, and social
history save events and populates the uuid mapping tables whenever those
records are saved.  This removes the populate uuid mechanisms from being
triggered on every single api requests and instead the uuids are
populated at the time of save.

Added events to the social history service that triggers whenever a
history data record is saved.

* Fix category bug with uuid mapping.

Had a bug which wasn't caught on my local unit tests but was on the
production ones (wierd) that the uuid mapping category was wrong.  It
was actually a pretty big flaw which is why I'm glad the unit test
caught it.  However the tests were in the wrong place since I moved
around where the mapping is done.  I created a new test for the events
subscriber, fixed the category problem for the mapping, and added an
additional test for the social history mapping which we didn't have.