Encounter Screen UI Improvements (#4741)
* Move all btn-primary classes to btn-secondary. btn-primary should only
be used if there is a single path the user is on. Additionally, use
btn-primary inside of forms to designate the Submit button, or the
action that occurs when the user presses enter. Most buttons in the
program should be secondary.
* Add a signature icon to better identify the eSign button.
* Make form title and author bigger and move that text to a heredoc
in preparation for a bigger twig migration.
* Improve Encounter Form Menu
* The Admin/Clinical/Misc buttons are now in a navbar, buttons have been
styled more appopriately, a new Twig filter is added.