Openemr fix #4888 document whitelist (#4913)
* Fixes #4888 Direct Document Whitelisting

I added an event to the sanitize.inc.php::isWhiteFile method called
the IsAcceptedFileFilterEvent that allows event consumers to modify the
whitelist on the fly or at the initial one time function invocation.

I put the event into the direct_message_check.inc::phimail_store
function that adds a listener to whitelist the direct mime type and then
removes the listener once the file has been uploaded.  This ensures that
no further php processing that happens after the function can sneek in a
file of the same mime-type into the system.

It looks like direct message files are already prefixed with a 'dm' in
the filename that's uploaded into OpenEMR so I'm not sure if we need to
do anything else more obvious to dennote this is a Direct file.  We
could stick the file into its own category but I'm not sure if that will
confuse users.

* Fix styles

* Fix style issue again.

Wish my styler matched the main project.  So bizarre.

* Fix namespace order and path prefixes

* Fix whitespace