Fixes #5006 User Default Billing Facility (#5007)
* Fixes #5006 User Default Billing Facility

We track it in the database but there wasn't a mechanism for setting a
provider's default billing facility. Since we need that for AMC and CCDA
I've gone ahead and added that in.

* Fix missing array key error, attribute quotes

Fixed on the add user field the missing array key error with the
billing_facility_id.  Also added quotes to the billing_facility_id
attribute for the selector.

Co-authored-by: Stephen Nielson <snielson@discoverandchange.com>