Fixes #5484 Vitals Form Show LOINC codes (#5485)
Made it so we visually show the loinc codes on the vital forms for
visual inspection of the data.  Since these are hard coded values on the
forms we need to visually show them as there is no list or database we
can demonstrate the codes to ONC.

A future improvement that would be nice would be to consolodate these
codes from both this file and the FhirObservationVitalsService class
which duplicates these codes.  In the interest of time we are leaving
them here.