Openemr fix #5496 vital form bugs (#5499)
* Fix vitals javascript errors.

kg to lb was throwing an error as we tried calling a Number method on a
string.

Vitals were saving in metric if you had metric set as the main unit and
entered in data that way.

Setting a vitals value to be empty was not emptying out the
corresponding conversion field and the saved input.

* Force vitals values to be floats

I haven't been able to reproduce this error on my version of php but on
some versions of php this was throwing a type conversion error.  We
parse as floats the values before we put them through number_format to
fix the type conversion problem.

* Fix some misnnamed variables.

* Fixes #5500 opt-out capability to template editor

Made it so double clicking the vitals form doesn't popup the dialog
template editor.