Multiple Patient Address Initial Support (#5396)
* Multiple Patient Address Initial Support

Fixes #5396 
Fixes #5502 
Fixes #5501 

Added a new data type for LBF to support Address List data type.

Addresses can be added with multiple addresses supported.  Initial
framework for saving addresses is here as well s retrieving them.

Rolled out the initial contact and contact address classes that can
be used to support Patients.  Decided to leverage the ORM classes
for Contact and ContactAddress to quickly persist and model
the data.

* Implement Multiple Address Save.

Implemented the multiple address save for contact and contact address.

Removed the duplicate code in ORDataObject to populate the object and
instead call the populateArray object.  This make sit possible to
overload the object and define custom behavior there.

Moved the Address, Contact, and ContactAddress class into our src folder
so we can follow normal PSR4 namespacing.

Fixed the primary key problem with the contact table in the database.

* PSR style fixes

* Add address type and use to address widget

* Address widget date selectors,county,reuse lbf

Made it so the address widget re-uses the lbf widget display code for
each of the address element codes.

Refactored the tables into divs to be mobile friendly.

Added in the date start and end dates so we can support previous
addresses.

Added the address use and address type selectors so people can identify
what kind of address it is.

Exposed the option js functions into the global window under the oeUI
container so we can have our add button on the state, country, use, and
type selector widgets.

* Fixes #5367 previous address support

Implemented the saving of the address list lbf template in the new
patient and the existing patient demographics page.

Added a new lbf field called additional_addresses to be displayed in the
contacts group layout.

Added the district field to support counties / districts in addresses.
Added the use and type FHIR data fields.
Also added a period start and period end to the addresses so we can
track the time period an address was in use.  New addresses default to a
current date of today with an empty end date.

Added new list options for address-uses and address-types to support the
new FHIR options.  The option_id is the FHIR key name.

Migrated some of the date formatting into the DateFormatterUtils object
so we can leverage them inside our objects without having to include the
formatter.inc.php file.

Added the address_list datatype to the restricted lbf fields so users
can't remove the datatype since we have to have it for USCDI support.

Refactored all of the styles of the address list widget to be responsive
and mobile friendly using bootstrap classes and styles.

Made the js functions of addresses consistent in naming conventions with
our other js functions.

Implemented the archiving / deactivating of the address list items.  Had
to switch from the BINARY(1) data type to a char value.  The
ORDataObject has a bug in that you can't save a value with a numeric 0
as its treated as empty.  I didn't want to try and change the logic and
mess up the other classes that extend ORDataObject, so in the interest
of time I switched isPrimary and status to be char fields that we key
off of.

Renamed the contact table fields to be consistent with fire naming
conventions per @mdsupport's recommendation.

Changed up the js events to be added in javacsript and not using inline
onclick events.  This simplifies the debugging immensely.

* Removed unused comments, added copyright notices

* Remove NULL from text columns

* Remove text display constraint,address list title

* Removed unused code, fixed event bug, label fixes

Had the close button not working when it was a new address, was only
working on existing address so had to fix that.

Also made the labels the same on the display as on the edit.

Removed unused code in the formatting.inc.php

* Fixes #5502,#5501 FHIR Addresses, Immunization

Fixes #5501 Immunization reason code column mispelling.
Fixes #5502 for multiple previous address support in FHIR.

Changed contact_address notes datatype to tinytext for review.

Added indexes to patient_history and contact_address to speed up
patient service search.

Added to api patient service the ability to include the patient
addresses. Had to refactor the queries as we were excluding any
previous names and previous addresses that didn't match the search
criteria.

Added to the /api/patient endpoint an address array of the patient data.
This includes both the main patient address and their additional address
information.  Eventually we can drop the inline address information in the
standard api.

Made the patient address search work on both the main address and additional
address information the patient has.

Made the DateFormatterUtils::dateStringToDateTime have support for times.