Mu3 3 author local cherry pick (#5588)
* Provenance Authorship for patient data

Add created_by and updated_by columns to patient_history, patient_data,
and history_data as well as modify dates and created dates so we can at
least track the last user who modified a piece of information and the
user that created a piece of information for provenance purposes.

A more detailed history can then be derived from the audit table log as
we'll be able to see what user modified the data on each patient record.

* Change author role code to be the ccda recommended

* ccda Local author provenance for problems.

Added in support to populate medical problems with the provenance
information from the lists table.

* ccda Provenance medical device, allergies

Made the medical device and allergies provenance information come from
the lists table with the last modifydate and username being the
provenance information.

* ccda medication provenance update

Made it so the prescriptions table keeps track of the user that created
the prescription and the user that last modified the prescription.

Updated provenance information to use the last modified date and the
last user that modified the prescription for the provenance information.
For historical data the assigned provider id is treated as the
provenance author.

Left the new table columns as nullable so we don't break any legacy
application use.

* ccda immunization provenance

Made the ccda use the provenance information from the immunizations
table instead of the encounter date.

* ccda provenance birth sex, social history

Made the birth sex social history provenance information point to the
patient_data table which is where gender is being pulled from.

Made the social history smoking, alcohol records pull from the
history_data table instead of from the encounters.

* ccda Provenance care team information

Made it so provenance care team populates from the patient_data table if
the primary care provider is set, or if the care coordination primary
provider is being overridden the care team is populated from the module
config settings.

* Fix allergy error for no known allergies

* Clinical notes provenance information

Make the assessment and clinical notes templates use the provenance
information from the form, and form_clinical_notes tables.

* ccda goals, health concerns, care plan provenance

Made goals, health concerns and care plan use the local provenance
information from the form_care_plan form.

* ccda provenance vitals update

Made the provenance information for the vitals come from the form_vitals
instead of the encounter

* Referral reason provenance ccda

* ccda provenance procedures

Made procedures use provenance from billing table instead of encounters

* ccda provenance for lab results

Made the provenance for lab results local instead of encounter

* Add provenance db changes to new database sql

Also escaped the column names in the upgrade file

* Fix ccda crashes w/ missing primary care provider

If the primary care provider was missing the ccda crashes.  This was
breaking on a new database install with a new patient who didn't have
the PCP set.

* Fixed undefined array key errors

Fixed a number of issues in the ccda generator where undefined array
keys were being thrown.  Re-tested on ETT and that all is working.

* Fix style issues.