Openemr fix 5665 vitals being squashed (#5668)
* Convert Vitals form from smarty to twig

Big overhaul of vitals from changing it from smarty to twig format.
Cleaned up a lot of the rendering as well so we can both easily extend
the vitals form and handle additional fields.  This is prep work to make
the form responsive.

Added some additional extension to twig so we could handle the wierd
date formats for the vital forms.

* Fix styles

* Fixes #5665 vitals form

Prevent the form from being squashed up and move the historical data
into its own section further down in the form.  Fix some minor twig
issues.

* Fix css styles

* Remove SMARTY controller class

Forgot to remove the smarty controller class now that we are no longer
using any of the smarty templates.